+++ packages/ace-config/src/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const envPaths = require('env-paths');
+const constants = require('./constants');
+const ConfigStore = require('./config-store');
+
+const paths = envPaths(constants.dirname, { suffix: null });
+
+const config = new ConfigStore({
+  cwd: paths.config,
+});
+
+module.exports = {
+  config,
+  constants: Object.freeze(constants),
+  paths,
+};