+++ packages/ace-core-legacy/bin/ace-http.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('@daisy/ace-http/bin/ace-http');