+++ packages/ace-core-legacy/src/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const ace = require('@daisy/ace-core');
+
+module.exports = ace;