+++ packages/ace-report/src/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const builders = require('./report-builders');
+const Report = require('./report');
+
+module.exports = {
+  builders,
+  Report,
+};