+++ packages/epub-utils/src/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const EPUB = require('./epub');
+
+module.exports = {
+  EPUB,
+};