+++ tests/runAceCLI.js
@@ -4,10 +4,8 @@
 
 const path = require('path');
 const spawn = require('cross-spawn');
-const babelCLIPkg = require('babel-cli/package');
 
-const BABEL_NODE = path.join(path.dirname(require.resolve('babel-cli')), babelCLIPkg.bin['babel-node']);
-const ACE_PATH = path.resolve(__dirname, '../src/cli/cli.js');
+const ACE_PATH = path.resolve(__dirname, '../packages/ace-cli/bin/ace.js');
 
 // return the result of the spawned process:
 //  [ 'status', 'signal', 'output', 'pid', 'stdout', 'stderr',
@@ -20,7 +18,7 @@ function ace(args, options = {}) {
     ? options.cwd
     : process.cwd();
 
-  const result = spawn.sync(BABEL_NODE, ['--', ACE_PATH].concat(args || []), { cwd, env });
+  const result = spawn.sync(ACE_PATH, args || [], { cwd, env });
 
   result.stdout = result.stdout && result.stdout.toString();
   result.stderr = result.stderr && result.stderr.toString();